+++ code/src/main.py
@@ -96,9 +96,9 @@ async def run_actor(actor_input: ActorInputsDb, payload: dict) -> None:
 
 async def get_embeddings(actor_input: ActorInputsDb) -> Embeddings:  # type: ignore[return]
     try:
-        Actor.log.info("Get embeddings class: %s", actor_input.embeddingsProvider.value)  # type: ignore[union-attr]
+        Actor.log.info("Get embeddings class: %s", actor_input.embeddingsProvider.value)
         embeddings = await get_embedding_provider(
-            actor_input.embeddingsProvider.value,  # type: ignore[union-attr]
+            str(actor_input.embeddingsProvider.value),
             actor_input.embeddingsApiKey,
             actor_input.embeddingsConfig,
         )