+++ code/src/vector_stores/milvus.py
@@ -29,7 +29,7 @@ class MilvusDatabase(Milvus, VectorDbBase):
     def dummy_vector(self) -> list[float]:
         if not self._dummy_vector and self.embeddings:
             self._dummy_vector = self.embeddings.embed_query("dummy")  # type: ignore
-        return self._dummy_vector  # type: ignore
+        return self._dummy_vector
 
     async def is_connected(self) -> bool:
         raise NotImplementedError